requirements.txt
forbiddenfruit
djangorestframework
pyyaml
uritemplate
cython
djangorestframework-simplejwt
psycopg2-binary
